requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
feedparser>=6.0.10
httpx>=0.25.0
beautifulsoup4>=4.12.0
python-dateutil>=2.8.2
APScheduler>=3.10.4
lxml>=4.9.3
